+++ R/assess_pb_bias_correction.R
@@ -57,7 +57,7 @@ assess_pb_bias_correction <- function(reference, gen_start_col, seed = 5,
                                       alle_freq_prior = list("const_scaled" = 1)) {
 
   # check that reference is formatted appropriately
-  check_refmix(reference, gen_start_col, "reference")
+  ploidies <- check_refmix(reference, gen_start_col, "reference")
 
   reference$collection <- factor(reference$collection, levels = unique(reference$collection))
   reference$repunit <- factor(reference$repunit, levels = unique(reference$repunit))
@@ -67,7 +67,8 @@ assess_pb_bias_correction <- function(reference, gen_start_col, seed = 5,
     dplyr::mutate(coll_int = as.integer(factor(reference$collection, levels = unique(reference$collection)))) %>%
     dplyr::select(repunit, coll_int) %>%
     dplyr::group_by(repunit, coll_int) %>%
-    dplyr::tally()
+    dplyr::tally() %>%
+    dplyr::filter(n > 0)
 
   if (is.null(reference$sample_type)) {
     sample_type <- rep("reference", nrow(reference))
@@ -78,7 +79,7 @@ assess_pb_bias_correction <- function(reference, gen_start_col, seed = 5,
   if (any(is.na(reference$repunit))) stop("repunit values may not be NAs" )
   if (any(is.na(reference$collection))) stop("collection values may not be NAs")
 
-  ref_params <- tcf2param_list(reference, gen_start_col, summ = F, alle_freq_prior = alle_freq_prior)
+  ref_params <- tcf2param_list(reference, gen_start_col, summ = F, alle_freq_prior = alle_freq_prior,  ploidies = ploidies)
 
   set.seed(seed)