+++ src/rcpp_per_locus_logls.cpp
@@ -31,6 +31,7 @@ NumericMatrix rcpp_per_locus_logls(List par_list) {
   IntegerVector A = as<IntegerVector>(par_list["A"]);
   IntegerVector CA = as<IntegerVector>(par_list["CA"]);
   IntegerVector coll = as<IntegerVector>(par_list["coll"]);
+  IntegerVector PLOID = as<IntegerVector>(par_list["ploidies"]);
   NumericVector DP = as<NumericVector>(par_list["DP"]);
   NumericVector sum_DP = as<NumericVector>(par_list["sum_DP"]);
   double gp;
@@ -40,7 +41,7 @@ NumericMatrix rcpp_per_locus_logls(List par_list) {
     LOO = 1;
     c = (coll[i] - 1);  // only compute logl for the population the individual is from
     for(l = 0; l < L; l++) {  // cycle over loci
-      GPROB_DIP(i, l, c, gp);
+      GPROB(i, l, c, gp);
       out(i, l) = log(gp);
     }
   }